Add a new style of test suite for the `witx` tooling similar to `*.wast` (#392)
* Add a wast-like testsuite for the witx tool

As the functionality of `witx` grows this will hopefully make adding
tests for new functionality as well as new kinds of tests easier. The
goal is to make it very easy to drop a test file with various directives
to exercise the functionality of `witx` and its internals.

For now the testsuite is quite simple, simply asserting whether
documents are either valid or invalid. My hope, though, is that this can
be expanded over time with more styles of assertions directives.

* Migrate union test to `union.witxt`

* Add documentation/roundtrip testing to all documents

* Translate multi-module test suite to `*.witxt`

* Move wasi tests into `witxt` test suites

* Convert representation tests to `*.witxt`

* Rebased onto origin/main