Adding functions for creating tensors from images
Two functions will be added, calculate_buffer_size
and convert_image. calculate_buffer_size will return
the number of bytes required to create the tensor.
convert_image will convert an image and store the
resulting tensor in the output buffer.

Image classification is a very common use case for
wasi-nn, this feature aims to make that easier and
faster.