Fix `get-output` return type
In porting the WITX definition to WIT, I inadvertently changed the
return type of `get-output`. @yamt identified this issue in #24--thank
you! This change fixes the return type to the original type,
`tensor-data`. (In the future, it may be interesting to return more
information about the tensor, but that can be a separate issue).