Re-add `*.witx` version of API
After some discussion with `wit-bindgen` maintainers, it is clear that
that tool is not yet ready to use for implementing WASI APIs inside
engines (e.g., Wasmtime). Since the existing Wasmtime infrastructure,
`wiggle`, uses the WITX syntax, this change re-adds the WITX version of
the wasi-nn API. This allows `wiggle` to continue to pull the WITX
specification directly from this repository until it is replaced by
`wit-bindgen`. All documentation has been removed from `wasi-nn.witx`
since we expect the WIT version to be the official version (at some
point, it would be nice to check in CI that the WITX and WIT versions expose
an identical API but this API does not do that).