ci: use new version of `wit-abi`
This change adds a new document `ml-example.md` which is generated from
the WIT world in `wasi-nn.wit`. `wasi-nn.wit` has to be moved back into
a `wit` directory to be found by the `wit-abi` tool to [check the
document]. Also, this updates documentation on how to keep
`ml-example.md` up to date.

[check the document]: https://github.com/WebAssembly/wit-abi-up-to-date/blob/71653db16a81a4319b87cada0bafff5a52f64a9b/main.js#L32