Remove `set-input`, `get-output`
As discussed in #43, there is no requirement to set up tensors prior to
calling `compute` as well as retrieving them separately afterwards. As
of #59, passing around tensors is cheap (they're resources now), so
there is no data copy necessary if we adopt this PR. This change
proposes removing the `set-input` and `get-output` functions, moving all
of the tensor-passing to `compute`. Closes #43.