Add properties to a `graph`
This follows up on a long-standing discussion from the [machine learning
working group]: a desire for some way to attach properties to an ML
`graph`. The various ML backends implementing this have mechanisms for
tweaking a model's behavior based on properties; this `string`-based
approach seems the most flexible way to access this.

[machine learning working group]: https://github.com/bytecodealliance/meetings/tree/main/machine-learning/