Remove error constructor (#76)
It should not be possible to construct an `error` from WebAssembly's
guest side. This isn't really useful to users (what can they do with
it?) and thus extends the API surface unnecessarily. This change removes
the `error` constructor, expecting that all wasi-nn errors will be
constructed on the host side.